Extract OnboardingDialog component
